+++ sdk/required_services.d.ts
@@ -34,16 +34,16 @@ export declare namespace google.ads.integration.service {
    * The collection of services required by the Ads app.
    */
   export interface RequiredServicesByAds {
-    authenticationService?:
+    authenticationService:
         authentication.ads.integration.service.partner.AuthenticationService;
 
-    businessService?: business.ads.integration.service.partner.BusinessService;
+    businessService: business.ads.integration.service.partner.BusinessService;
 
-    conversionTrackingService?: partnerConversionTracking.ads.integration
-        .service.partner.ConversionTrackingService;
+    conversionTrackingService: partnerConversionTracking.ads.integration.service
+        .partner.ConversionTrackingService;
 
-    termsAndConditionsService?: partnerTermsAndConditions.ads.integration
-        .service.partner.TermsAndConditionsService;
+    termsAndConditionsService: partnerTermsAndConditions.ads.integration.service
+        .partner.TermsAndConditionsService;
 
     accountStatusService?:
         accountStatus.ads.integration.service.partner.AccountStatusService;